+++ packages/hoc-react/src/onlyCsr.tsx
@@ -6,13 +6,13 @@ import { SProps } from 'ssr-types'
 type FC<T={}, U={}> = (props: Partial<SProps<T>> & Partial<U>) => JSX.Element
 
 function onlyCsr<T={}, U={}> (WrappedComponent: FC<T, U>) {
-  return (props: SProps<T> & U) => {
+  return (props: Partial<SProps<T>> & U) => {
     const [isClient, setIsClient] = useState(false)
     useEffect(() => {
       setIsClient(true)
     }, [])
     return (
-      isClient ? <WrappedComponent {...props}></WrappedComponent> : React.Fragment
+      isClient ? <WrappedComponent {...props}></WrappedComponent> : <></>
     )
   }
 }