+++ packages/plugin-react/src/config/base.ts
@@ -1,16 +1,42 @@
 
 import { join } from 'path'
 import { Mode } from 'ssr-types'
-import { getCwd, loadConfig, setStyle, addImageChain, loadModuleFromFramework } from 'ssr-common-utils'
+import { getCwd, loadConfig, setStyle, addImageChain, loadModuleFromFramework, getPkgJson } from 'ssr-common-utils'
 import * as WebpackChain from 'webpack-chain'
 import * as webpack from 'webpack'
+import { coerce } from 'semver'
 
 const MiniCssExtractPlugin = require(loadModuleFromFramework('ssr-mini-css-extract-plugin'))
 const WebpackBar = require('webpackbar')
 const loadModule = loadModuleFromFramework
 
+const antdVersion = getPkgJson().dependencies?.['antd'] ?? getPkgJson().devDependencies?.['antd']
+const isAntd4 = coerce(antdVersion)?.major === 4
 const addBabelLoader = (chain: WebpackChain.Rule<WebpackChain.Module>, envOptions: any, isServer: boolean) => {
   const { babelOptions, isDev } = loadConfig()
+  const plugins = [
+    [loadModule('@babel/plugin-transform-runtime'), {
+      regenerator: false,
+      corejs: false,
+      helpers: true
+    }],
+    [loadModule('@babel/plugin-proposal-private-methods'), { loose: true }],
+    [loadModule('@babel/plugin-proposal-private-property-in-object'), { loose: true }],
+    ...babelOptions?.plugins ?? []
+  ]
+  if (isAntd4) {
+    plugins.push([
+      loadModule('babel-plugin-import'),
+      {
+        libraryName: 'antd',
+        libraryDirectory: 'lib',
+        style: true
+      }, 'antd'
+    ])
+  }
+  if (!isServer && isDev) {
+    plugins.push(loadModule('react-refresh/babel'))
+  }
   chain.use('babel-loader')
     .loader(loadModule('babel-loader'))
     .options({
@@ -22,33 +48,20 @@ const addBabelLoader = (chain: WebpackChain.Rule<WebpackChain.Module>, envOption
           loadModule('@babel/preset-env'),
           envOptions
         ],
-        [loadModule('babel-preset-react-app'), { flow: false, typescript: true }],
+        [loadModule('babel-preset-react-app'), {
+          flow: false,
+          typescript: true,
+          runtime: 'automatic'
+        }],
         ...babelOptions?.presets ?? []
       ],
-      plugins: [
-        [loadModule('@babel/plugin-transform-runtime'), {
-          regenerator: false,
-          corejs: false,
-          helpers: true
-        }],
-        [
-          loadModule('babel-plugin-import'),
-          {
-            libraryName: 'antd',
-            libraryDirectory: 'lib',
-            style: true
-          }, 'antd'
-        ],
-        [loadModule('@babel/plugin-proposal-private-methods'), { loose: true }],
-        [loadModule('@babel/plugin-proposal-private-property-in-object'), { loose: true }],
-        ...babelOptions?.plugins ?? []
-      ].concat((!isServer && isDev) ? loadModule('react-refresh/babel') : [])
+      plugins: plugins
     })
     .end()
 }
 const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
   const config = loadConfig()
-  const { moduleFileExtensions, useHash, chainBaseConfig, corejsOptions, babelExtraModule, alias, define } = config
+  const { moduleFileExtensions, useHash, chainBaseConfig, corejsOptions, babelExtraModule, alias, define, babelOptions } = config
   const mode = process.env.NODE_ENV as Mode
   const envOptions = {
     modules: false,
@@ -72,7 +85,6 @@ const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
     chain.resolve.alias
       .set(item, alias[item])
   })
-
   addImageChain(chain, isServer)
 
   const babelModule = chain.module
@@ -80,6 +92,7 @@ const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
     .test(/\.(js|mjs|jsx|ts|tsx)$/)
     .exclude
     .add(/node_modules|core-js/)
+    .add(babelOptions?.exclude as Array<string|RegExp> ?? [])
     .end()
 
   chain.module
@@ -94,12 +107,7 @@ const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
     .include
     .add([/ssr-plugin-react/, /ssr-client-utils/, /ssr-hoc-react/, /ssr-common-utils/])
 
-  let babelForExtraModule
-  if (babelExtraModule) {
-    babelForExtraModule = module.add(babelExtraModule).end().exclude.add(/core-js/).end()
-  } else {
-    babelForExtraModule = module.end().exclude.add(/core-js/).end()
-  }
+  const babelForExtraModule = module.add(babelExtraModule ?? []).add(babelOptions?.include as Array<string|RegExp> ?? []).end().exclude.add(/core-js/).end()
 
   addBabelLoader(babelModule, envOptions, isServer)
   addBabelLoader(babelForExtraModule, envOptions, isServer)