+++ packages/plugin-vue/src/config/base.ts
@@ -61,7 +61,7 @@ const addBabelLoader = (chain: WebpackChain.Rule<WebpackChain.Module>, envOption
 
 const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
   const config = loadConfig()
-  const { moduleFileExtensions, useHash, chainBaseConfig, corejsOptions, ssrVueLoaderOptions, csrVueLoaderOptions, babelExtraModule, alias, define } = config
+  const { moduleFileExtensions, useHash, chainBaseConfig, corejsOptions, ssrVueLoaderOptions, csrVueLoaderOptions, babelExtraModule, alias, define, babelOptions } = config
 
   let vueLoaderOptions = {
     babelParserPlugins: ['jsx', 'classProperties', 'decorators-legacy']
@@ -126,6 +126,7 @@ const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
     .test(/\.(js|mjs|jsx|ts|tsx)$/)
     .exclude
     .add(/node_modules|core-js/)
+    .add(babelOptions?.exclude as Array<string|RegExp> ?? [])
     .end()
 
   const module = chain.module
@@ -134,12 +135,7 @@ const getBaseConfig = (chain: WebpackChain, isServer: boolean) => {
     .include
     .add([/ssr-plugin-vue/, /ssr-client-utils/, /ssr-common-utils/])
 
-  let babelForExtraModule
-  if (babelExtraModule) {
-    babelForExtraModule = module.add(babelExtraModule).end().exclude.add(/core-js/).end()
-  } else {
-    babelForExtraModule = module.end().exclude.add(/core-js/).end()
-  }
+  const babelForExtraModule = module.add(babelExtraModule ?? []).add(babelOptions?.include as Array<string|RegExp> ?? []).end().exclude.add(/core-js/).end()
 
   addBabelLoader(babelModule, envOptions)
   addBabelLoader(babelForExtraModule, envOptions)