+++ packages/plugin-vue/src/entry/server-entry.ts
@@ -15,7 +15,8 @@ const serverRender = async (ctx: ISSRContext, config: IConfig) => {
   const router = createRouter()
   const store = createStore()
   setStore(store)
-  const [path, url] = [normalizePath(ctx.request.path, prefix), normalizePath(ctx.request.url, prefix)]
+  const rawPath = ctx.request.path ?? ctx.request.routerPath
+  const [path, url] = [normalizePath(rawPath, prefix), normalizePath(ctx.request.url, prefix)]
   const routeItem = findRoute<IFeRouteItem>(FeRoutes, path)
 
   if (!routeItem) {
@@ -89,12 +90,12 @@ const serverRender = async (ctx: ISSRContext, config: IConfig) => {
           id: 'app'
         }
       }, [h(App, {
-        props: { ctx, config, fetchData: combineAysncData, reactiveFetchData: { value: combineAysncData } }
+        props: { ctx, config, fetchData: combineAysncData, asyncData: { value: combineAysncData }, reactiveFetchData: { value: combineAysncData } }
       })])
       return h(
         Layout,
         {
-          props: { ctx, config, asyncData: combineAysncData, fetchData: layoutFetchData, reactiveFetchData: { value: layoutFetchData } }
+          props: { ctx, config, asyncData: Object.assign(combineAysncData, { value: combineAysncData }), fetchData: layoutFetchData, reactiveFetchData: { value: layoutFetchData } }
         },
         [
           h('template', {